NN-2923: Break out API and React routing to correctly enable CSRF
